fix: (Platform) approval flow stackblitz fixes, angular without Ivy enabled support
