+++ apps/docs/src/app/platform/component-docs/platform-approval-flow/platform-approval-flow-docs.component.html
@@ -1,5 +1,5 @@
 <h2>Approval Flow</h2>
 <component-example>
-    <fdp-approval-flow-example></fdp-approval-flow-example>
+    <fdp-platform-approval-flow-example></fdp-platform-approval-flow-example>
 </component-example>
 <code-example [exampleFiles]="basic"></code-example>
