+++ apps/docs/src/app/platform/component-docs/platform-approval-flow/platform-approval-flow-examples/platform-approval-flow-example.component.html
@@ -1,0 +1,123 @@
+<fdp-approval-flow
+    title='Basic Approval Flow Demo'
+    [dataSource]='dataSource'
+    [userDetailsTemplate]='userDetailsTemplate'
+    [checkDueDate]='checkDueDate'
+    [allowSendRemindersForStatuses]='sendReminderStatuses'
+    [isEditAvailable]='editModeEnabled'
+    (nodeClick)='nodeClick($event)'
+>
+</fdp-approval-flow>
+
+<ng-template #userDetailsTemplate let-data='data'>
+    <h4 style='margin-bottom: 1rem;'>Contact info</h4>
+
+    <p style='display: flex; flex-direction: column;'>
+        <span>Mobile</span>
+        <a href='javascript:void(0)'>{{ data?.phone }}</a>
+    </p>
+
+    <p style='display: flex; flex-direction: column;'>
+        <span>Phone</span>
+        <a href='javascript:void(0)'>{{ data?.phone }}</a>
+    </p>
+
+    <p style='display: flex; flex-direction: column;'>
+        <span>E-mail</span>
+        <a href='javascript:void(0)'>{{ data?.email }}</a>
+    </p>
+
+    <h4 style='margin-top: 1.5rem; margin-bottom: 1rem;'>Company</h4>
+
+    <p style='display: flex; flex-direction: column;'>
+        <span>Name</span>
+        <a href='javascript:void(0)'>Company A{{ data?.email }}</a>
+    </p>
+
+    <p style='display: flex; flex-direction: column;'>
+        <span>Address</span>
+        <a href='javascript:void(0)'>481 West Street, Anytown OH, 83749, USA</a>
+    </p>
+</ng-template>
+
+<!-- Playground code below, not related to the component itself -->
+
+<p>
+    Selected example:
+    <select [(ngModel)]='selectedExample' (ngModelChange)='dataSource.selectGraph(selectedExample)'>
+        <option *ngFor='let example of examples' [value]='example'>{{ example | titlecase }}</option>
+    </select>
+</p>
+<p>Enable "Edit mode": <input type='checkbox' [(ngModel)]='editModeEnabled'></p>
+<p>Enable due date warnings: <input type='checkbox' [(ngModel)]='checkDueDate'></p>
+<p>Set all statuses to "Not Started": <input type='checkbox' [(ngModel)]='setNotStartedStatuses' (ngModelChange)='setNotStarted()'>
+</p>
+<p>
+    Disable all node actions:
+    <input type='checkbox' [(ngModel)]='nodeActionsDisabled' (ngModelChange)='toggleNodeActions($event)'>
+</p>
+<p>Disable certain node actions: </p>
+<div>
+    <p>
+        <input
+            id='disableAddingNodeBefore'
+            [disabled]='nodeActionsDisabled'
+            type='checkbox'
+            [(ngModel)]='nodeActionsConfig.disableAddBefore'
+            (ngModelChange)='toggleSpecificNodeAction("disableAddBefore", $event)'
+        >
+        <label for='disableAddingNodeBefore'>Disable adding before</label>
+    </p>
+
+    <p>
+        <input
+            id='disableAddingNodeAfter'
+            [disabled]='nodeActionsDisabled'
+            type='checkbox'
+            [(ngModel)]='nodeActionsConfig.disableAddAfter'
+            (ngModelChange)='toggleSpecificNodeAction("disableAddAfter", $event)'
+        >
+        <label for='disableAddingNodeAfter'>Disable adding after</label>
+    </p>
+
+    <p>
+        <input
+            id='disableAddingNodeParallel'
+            [disabled]='nodeActionsDisabled'
+            type='checkbox'
+            [(ngModel)]='nodeActionsConfig.disableAddParallel'
+            (ngModelChange)='toggleSpecificNodeAction("disableAddParallel", $event)'
+        >
+        <label for='disableAddingNodeParallel'>Disable adding parallel</label>
+    </p>
+
+    <p>
+        <input
+            id='disableEditingNode'
+            [disabled]='nodeActionsDisabled'
+            type='checkbox'
+            [(ngModel)]='nodeActionsConfig.disableEdit'
+            (ngModelChange)='toggleSpecificNodeAction("disableEdit", $event)'
+        >
+        <label for='disableEditingNode'>Disable editing</label>
+    </p>
+
+    <p>
+        <input
+            id='disableRemovingNode'
+            [disabled]='nodeActionsDisabled'
+            type='checkbox'
+            [(ngModel)]='nodeActionsConfig.disableRemove'
+            (ngModelChange)='toggleSpecificNodeAction("disableRemove", $event)'
+        >
+        <label for='disableRemovingNode'>Disable removing</label>
+    </p>
+</div>
+<p style='display: flex;align-items: center;'>Allow sending reminders to approvers with statuses:
+    <fd-multi-input
+        style='margin-left: .5rem;'
+        [compact]='true'
+        [dropdownValues]='allStatuses'
+        [(ngModel)]='sendReminderStatuses'
+    ></fd-multi-input>
+</p>
